feat: add common exports for otlp-transformer
